refactor: index.ts에 api 연결 코드 모음
